gendatafun {COWA}	R Documentation

Generate the simulated data

Description

Generate the simulated data based on multiple list experiments.

Usage



  gendatafun(v,T,d,m,theta0)


Arguments

v
an integer, population size of each primary period

T
an integer, the number of primary period

d
an integer, the number of agencies(lists)

m
an integer, the dimension of covariates

theta0
an vector, the initial value

Details

nothing!

Value

return a list whose elements including: list with covariates of each primary period, list with multilist observe of each primary period, vector with observed numbers of each primary period, initial vector value

Note

nothing

Author(s)

Li

Examples


  rm(list=ls(all=TRUE))
  library(MASS)
  v=1000
  T=5
  d=4
  m<-5
  set.seed(1234)
  theta0 <-round(rnorm((T+m*d),0,0.8),1)
  DATA <-gendatafun(v,T,d,m)


[Package COWA version 0.1.0 Index]